Server/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
numpy==1.19.5
Pillow==9.3.0
Pillow==10.3.0
protobuf==3.20.0
python-dotenv==1.0.1
tensorflow==2.6.0
